+++ autd3-driver/src/datagram/silencer/mod.rs
@@ -18,7 +18,7 @@ pub type SilencerFixedCompletionTime = Silencer<FixedCompletionTime>;
 pub type SilencerFixedUpdateRate = Silencer<FixedUpdateRate>;
 
 impl Silencer<()> {
-    pub const fn fixed_update_rate(
+    pub const fn from_update_rate(
         update_rate_intensity: u16,
         update_rate_phase: u16,
     ) -> Silencer<FixedUpdateRate> {
@@ -30,7 +30,7 @@ impl Silencer<()> {
         }
     }
 
-    pub const fn fixed_completion_steps(
+    pub const fn from_completion_steps(
         steps_intensity: u16,
         steps_phase: u16,
     ) -> Silencer<FixedCompletionSteps> {
@@ -43,7 +43,7 @@ impl Silencer<()> {
         }
     }
 
-    pub const fn fixed_completion_time(
+    pub const fn from_completion_time(
         time_intensity: std::time::Duration,
         time_phase: std::time::Duration,
     ) -> Silencer<FixedCompletionTime> {
@@ -80,31 +80,31 @@ mod tests {
     }
 
     #[test]
-    fn fixed_update_rate() {
-        let s = Silencer::fixed_update_rate(1, 2);
+    fn from_update_rate() {
+        let s = Silencer::from_update_rate(1, 2);
         assert_eq!(s.update_rate_intensity(), 1);
         assert_eq!(s.update_rate_phase(), 2);
     }
 
     #[test]
-    fn fixed_completion_steps_mul() {
-        let s = Silencer::fixed_completion_steps(1, 1);
+    fn from_completion_steps_mul() {
+        let s = Silencer::from_completion_steps(1, 1);
         let s = s * 2;
         assert_eq!(s.completion_steps_intensity(), 2);
         assert_eq!(s.completion_steps_phase(), 2);
     }
 
     #[test]
-    fn fixed_completion_steps_div() {
-        let s = Silencer::fixed_completion_steps(2, 2);
+    fn from_completion_steps_div() {
+        let s = Silencer::from_completion_steps(2, 2);
         let s = s / 2;
         assert_eq!(s.completion_steps_intensity(), 1);
         assert_eq!(s.completion_steps_phase(), 1);
     }
 
     #[test]
-    fn fixed_completion_time() {
-        let s = Silencer::fixed_completion_time(
+    fn from_completion_time() {
+        let s = Silencer::from_completion_time(
             std::time::Duration::from_secs(1),
             std::time::Duration::from_secs(1),
         );
@@ -116,8 +116,8 @@ mod tests {
     }
 
     #[test]
-    fn fixed_completion_time_mul() {
-        let s = Silencer::fixed_completion_time(
+    fn from_completion_time_mul() {
+        let s = Silencer::from_completion_time(
             std::time::Duration::from_secs(1),
             std::time::Duration::from_secs(1),
         );
@@ -130,8 +130,8 @@ mod tests {
     }
 
     #[test]
-    fn fixed_completion_time_div() {
-        let s = Silencer::fixed_completion_time(
+    fn from_completion_time_div() {
+        let s = Silencer::from_completion_time(
             std::time::Duration::from_secs(2),
             std::time::Duration::from_secs(2),
         );