+++ autd3-firmware-emulator/tests/op/modulation.rs
@@ -174,7 +174,7 @@ fn mod_freq_div_too_small() -> anyhow::Result<()> {
         .with_segment(Segment::S0, Some(TransitionMode::Immediate));
         assert_eq!(Ok(()), send(&mut cpu, d, &geometry, &mut tx));
 
-        let d = Silencer::fixed_completion_steps(
+        let d = Silencer::from_completion_steps(
             SILENCER_STEPS_INTENSITY_DEFAULT,
             SILENCER_STEPS_PHASE_DEFAULT,
         );
@@ -191,7 +191,7 @@ fn mod_freq_div_too_small() -> anyhow::Result<()> {
         .with_segment(Segment::S1, None);
         assert_eq!(Ok(()), send(&mut cpu, d, &geometry, &mut tx));
 
-        let d = Silencer::fixed_completion_steps(
+        let d = Silencer::from_completion_steps(
             SILENCER_STEPS_PHASE_DEFAULT * 2,
             SILENCER_STEPS_PHASE_DEFAULT,
         );